# Conflicts:
#	backend/src/main/java/com/amigoscode/Main.java
#	backend/src/main/java/com/amigoscode/customer/Customer.java
#	backend/src/main/java/com/amigoscode/customer/CustomerDTO.java
#	backend/src/main/java/com/amigoscode/customer/CustomerJDBCDataAccessService.java
#	backend/src/main/java/com/amigoscode/customer/CustomerJPADataAccessService.java
#	backend/src/main/java/com/amigoscode/customer/CustomerListDataAccessService.java
#	backend/src/main/java/com/amigoscode/customer/CustomerRepository.java
#	backend/src/main/java/com/amigoscode/customer/CustomerRowMapper.java
#	backend/src/main/java/com/amigoscode/customer/CustomerService.java
#	backend/src/main/java/com/amigoscode/s3/S3Buckets.java
#	backend/src/main/java/com/amigoscode/s3/S3Config.java
#	backend/src/main/java/com/amigoscode/s3/S3Service.java
#	backend/src/main/resources/application.yml
#	backend/src/main/resources/db/migration/V2__Add_Customer_Profile_Image.sql
#	backend/src/test/java/com/amigoscode/customer/CustomerRowMapperTest.java
#	backend/src/test/java/com/amigoscode/customer/CustomerServiceTest.java
#	backend/src/test/java/com/amigoscode/journey/CustomerIT.java

+++ backend/src/main/java/com/amigoscode/customer/CustomerService.java
@@ -5,7 +5,6 @@ import com.amigoscode.exception.RequestValidationException;
 import com.amigoscode.exception.ResourceNotFoundException;
 import com.amigoscode.s3.S3Buckets;
 import com.amigoscode.s3.S3Service;
-import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
@@ -23,13 +22,12 @@ public class CustomerService {
     private final CustomerDTOMapper customerDTOMapper;
     private final PasswordEncoder passwordEncoder;
     private final S3Service s3Service;
+
     private final S3Buckets s3Buckets;
 
     public CustomerService(@Qualifier("jdbc") CustomerDao customerDao,
                            CustomerDTOMapper customerDTOMapper,
-                           PasswordEncoder passwordEncoder,
-                           S3Service s3Service,
-                           S3Buckets s3Buckets) {
+                           PasswordEncoder passwordEncoder, S3Service s3Service, S3Buckets s3Buckets) {
         this.customerDao = customerDao;
         this.customerDTOMapper = customerDTOMapper;
         this.passwordEncoder = passwordEncoder;
@@ -73,16 +71,13 @@ public class CustomerService {
     }
 
     public void deleteCustomerById(Integer customerId) {
-        checkIfCustomerExistsOrThrow(customerId);
-        customerDao.deleteCustomerById(customerId);
-    }
-
-    private void checkIfCustomerExistsOrThrow(Integer customerId) {
         if (!customerDao.existsCustomerById(customerId)) {
             throw new ResourceNotFoundException(
                     "customer with id [%s] not found".formatted(customerId)
             );
         }
+
+        customerDao.deleteCustomerById(customerId);
     }
 
     public void updateCustomer(Integer customerId,
@@ -121,10 +116,12 @@ public class CustomerService {
 
         customerDao.updateCustomer(customer);
     }
-
-    public void uploadCustomerProfileImage(Integer customerId,
-                                           MultipartFile file) {
-        checkIfCustomerExistsOrThrow(customerId);
+    public void uploadCustomerProfileImage(Integer customerId, MultipartFile file){
+        if (!customerDao.existsCustomerById(customerId)) {
+            throw new ResourceNotFoundException(
+                    "customer with id [%s] not found".formatted(customerId)
+            );
+        }
         String profileImageId = UUID.randomUUID().toString();
         try {
             s3Service.putObject(
@@ -133,10 +130,11 @@ public class CustomerService {
                     file.getBytes()
             );
         } catch (IOException e) {
-            throw new RuntimeException("failed to upload profile image", e);
+            throw new RuntimeException(e);
         }
         customerDao.updateCustomerProfileImageId(profileImageId, customerId);
     }
+
 
     public byte[] getCustomerProfileImage(Integer customerId) {
         var customer = customerDao.selectCustomerById(customerId)
@@ -145,16 +143,14 @@ public class CustomerService {
                         "customer with id [%s] not found".formatted(customerId)
                 ));
 
-        if (StringUtils.isBlank(customer.profileImageId())) {
+        if (customer.profileImageId().isBlank()) {
             throw new ResourceNotFoundException(
-                    "customer with id [%s] profile image not found".formatted(customerId));
+                    "customer with id [%s]  profile image not found".formatted(customerId));
         }
-
-        byte[] profileImage = s3Service.getObject(
+        return s3Service.getObject(
                 s3Buckets.getCustomer(),
                 "profile-images/%s/%s".formatted(customerId, customer.profileImageId())
         );
-        return profileImage;
     }
 }
